+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Build Parameters Gradle plugin - Changelog
 
+## Version 1.3
+* [New] [#18](https://github.com/gradlex-org/build-parameters/issues/18) Fail the build when it's running on an unsupported Gradle version
+
 ## Version 1.2
 * [New] [#42](https://github.com/gradlex-org/build-parameters/issues/42) Boolean parameters: empty string maps to 'true' and invalid value fails the build (instead of silently mapping to 'false')
 * [New] [#40](https://github.com/gradlex-org/build-parameters/issues/40) Allow defining parameters without configuration action